moved notfound component to pages
